requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
geographiclib==2.1
geopy==2.4.1
idna==3.11
numpy==2.4.4
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.0
six==1.17.0
tqdm==4.67.3
urllib3==2.6.3
